Extact shared photo code
