+++ app/models/messages/cyclestreets_photo_message.rb
@@ -1,2 +1,10 @@
 class CyclestreetsPhotoMessage < MessageComponent
+  include Photo
+  include Locatable
+
+  validates :location, :photo, presence: true
+
+  def searchable_text
+    [caption, cyclestreet_caption, cyclestreet_category].join(' ')
+  end
 end
