+++ spec/models/cyclestreets_photo_spec.rb
@@ -1,5 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe CyclestreetsPhoto, type: :model do
-  pending "add some examples to (or delete) #{__FILE__}"
+describe CyclestreetsPhoto, type: :model do
+  it { is_expected.to validate_presence_of(:photo) }
+  it { is_expected.to validate_presence_of(:location) }
 end
